Added the ability to keep track of rooms
After this commit, the server has the capability to keep track of multiple rooms (2+) with each having multiple rangers (2+).
In addition, the python client now knows about who the other rangers are in the room. When the a user disconnects from the server/room, it will remove them from the dictionary keeping track of that information.

Python client also sends coordinates of it's current ranger to the server. Not yet done implementing.